+++ src/i18n-l10n-editor/models/ije-view.ts
@@ -1,6 +1,7 @@
 export interface IJEView {
     type: IJEViewType;
     selectionId?: number;
+    selectionFolder?: string;
 }
 export enum IJEViewType {
     TABLE = 'table',